requirements.txt
Django>=5.2,<5.3
django-tenants>=3.9.0
djangorestframework>=3.14
celery>=5.3
redis>=5.0
requests>=2.31
python-dotenv>=1.0
pillow>=10.0
whitenoise>=6.6
psycopg2-binary>=2.9
openpyxl>=3.1
pdfplumber>=0.11
python-dateutil>=2.8
